Update fb_modules.py
the popupwindow should be in a "try/pass" otherwise it may give an error for the whole selectionentered module. 
It might give an error when there is a question but no answer, which should still be allowed as you only need 1 question card, for example when you take just notes but not necessarily make Q&A cards.